requirements.txt
cachelib==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-MySQLdb==0.2.0
Flask-Session==0.4.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
Pillow==8.3.2
protobuf==3.17.3
pytesseract==0.3.8
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
